requirements.txt
EdgeGPT==0.11.6
httpx==0.24.1
nonebot==1.9.1
nonebot_plugin_guild_patch==0.2.3
pydantic==1.10.7
revChatGPT==6.4.4
tencentcloud_sdk_python==3.0.617
